Alterações feitas no exercício 01

+++ src/exercicio1oo/Circulo.java
@@ -1,18 +1,22 @@
 package exercicio1oo;
 
 public class Circulo {
-    public float raio;
-    public Circulo(float raio){
-        this.raio=raio;
+    public double raio;
+
+    public Circulo(double raio) {
+        this.raio = raio;
     }
-    public float calcularDiametro(){
+
+    public double calcularDiametro() {
         return 2 * raio;
     }
-    public float calcularPerimetro(){
-        return 2*(float)Math.PI*raio;
+
+    public double calcularPerimetro() {
+        return 2 * (double) Math.PI * raio;
     }
-    public float calcularArea(){
-        return (float)Math.PI*raio*raio;
+
+    public double calcularArea() {
+        return (double) Math.PI * raio * raio;
     }
 }
